"Finish" the demo
