+++ src/widget/compose_style/padding.rs
@@ -1,81 +1,19 @@
-use glazier::kurbo::{Affine, Size};
+use glazier::kurbo::{Affine, Point, Size};
 
-use crate::{id::Id, widget::Pod, View, Widget};
-
-pub fn padding<V: View<T, A>, T, A>(width: f64, view: V) -> impl View<T, A>
-where
-    V::Element: 'static,
-{
-    PaddingView::new(width, view)
-}
-
-pub struct PaddingView<V> {
-    width: f64,
-    view: V,
-}
-
-impl<T, A, V: View<T, A>> View<T, A> for PaddingView<V>
-where
-    V::Element: 'static,
-{
-    type State = (Id, V::State);
-
-    type Element = PaddingWidget;
-
-    fn build(&self, cx: &mut crate::view::Cx) -> (crate::id::Id, Self::State, Self::Element) {
-        let (id, (child_id, state, element)) = cx.with_new_id(|cx| self.view.build(cx));
-        let element = PaddingWidget::new(Pod::new(element), self.width);
-        (id, (child_id, state), element)
-    }
-
-    fn rebuild(
-        &self,
-        cx: &mut crate::view::Cx,
-        prev: &Self,
-        id: &mut crate::id::Id,
-        (child_id, state): &mut Self::State,
-        element: &mut Self::Element,
-    ) -> bool {
-        let mut changed = prev.width != self.width;
-        cx.with_id(*id, |cx| {
-            let child_element = element.widget.downcast_mut().unwrap();
-            changed |= self
-                .view
-                .rebuild(cx, &prev.view, child_id, state, child_element)
-        });
-        changed
-    }
-
-    fn event(
-        &self,
-        id_path: &[crate::id::Id],
-        (child_id, state): &mut Self::State,
-        event: Box<dyn std::any::Any>,
-        app_state: &mut T,
-    ) -> crate::event::EventResult<A> {
-        let (left, right) = id_path.split_at(1);
-        assert!(left[0] == *child_id);
-        self.view.event(right, state, event, app_state)
-    }
-}
-
-impl<V> PaddingView<V> {
-    fn new(width: f64, view: V) -> Self {
-        PaddingView { width, view }
-    }
-}
+use crate::{widget::Pod, View, Widget};
 
 pub struct PaddingWidget {
-    widget: Pod,
+    pub(crate) widget: Pod,
     width: f64,
 }
 
 impl PaddingWidget {
     pub fn new(widget: Pod, width: f64) -> Self {
-        Self {
-            widget: widget,
-            width,
-        }
+        Self { widget, width }
+    }
+
+    pub fn set_width(&mut self, width: f64) {
+        self.width = width;
     }
 }
 
@@ -93,6 +31,7 @@ impl Widget for PaddingWidget {
     }
 
     fn update(&mut self, cx: &mut crate::widget::UpdateCx) {
+        cx.request_layout();
         self.widget.update(cx)
     }
 
@@ -111,7 +50,12 @@ impl Widget for PaddingWidget {
         proposed_size: glazier::kurbo::Size,
     ) -> glazier::kurbo::Size {
         let padding_size = Size::new(self.width * 2., self.width * 2.);
-        self.widget.layout(cx, proposed_size - padding_size) + padding_size
+        let result = self.widget.layout(cx, proposed_size - padding_size) + padding_size;
+        self.widget.state.origin = Point {
+            x: self.width,
+            y: self.width,
+        };
+        result
     }
 
     fn paint(&mut self, cx: &mut crate::widget::PaintCx, builder: &mut piet_scene::SceneBuilder) {
